% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBoard.R
\name{generateBoard}
\alias{generateBoard}
\title{Generate a Minesweeper game board matrix}
\usage{
generateBoard(nrows, ncols, nmines)
}
\arguments{
\item{nrows}{The number of rows in the game board.}

\item{ncols}{The number of columns in the game board.}

\item{nmines}{The number of mines to be placed on the game board.}
}
\value{
A nrows x ncols matrix representing the Minesweeper game board.
}
\description{
This function generates a Minesweeper game board matrix with a specified number of rows, columns, and mines.
The generated matrix contains -1 for cells with mines and the number of adjacent mines for other cells.
}
\examples{
board <- generateBoard(5, 5, 10)
}
