% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCell.R
\name{openCell}
\alias{openCell}
\title{Open a cell and flood-fill adjacent cells in a Minesweeper game}
\usage{
openCell(r, c, gameState)
}
\arguments{
\item{r}{The row index of the cell to open.}

\item{c}{The column index of the cell to open.}

\item{gameState}{A list representing the current game state, containing the following elements:
\itemize{
\item board: A matrix representing the minesweeper board. -1 indicates a mine, other values indicate the number of adjacent mines.
\item opened: A matrix representing whether a cell has been opened (TRUE) or not (FALSE).
\item nrows: The number of rows in the game board.
\item ncols: The number of columns in the game board.
\item gameOver: A boolean indicating whether the game is over (TRUE) or not (FALSE).
}}
}
\value{
A modified gameState list with the specified cell and its adjacent cells opened.
}
\description{
This function recursively opens a cell and its adjacent cells in a Minesweeper game state.
If the cell contains a mine, the game is marked as over. If the cell is empty (has no adjacent mines),
the function will recursively open all its adjacent cells.
}
\examples{
gameState <- list(board = matrix(0, 3, 3),
opened = matrix(FALSE, 3, 3),
nrows = 3,
ncols = 3,
gameOver = FALSE)
newGameState <- openCell(1, 1, gameState)
}
