% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweeper.R
\name{sweeper}
\alias{sweeper}
\title{Create a Minesweeper game using Shiny}
\usage{
sweeper()
}
\value{
A Shiny app object for the Minesweeper game.
}
\description{
This function creates a Minesweeper game using the Shiny web framework. The game board is dynamically
generated based on user input for the number of rows, columns, and mines. Users can open cells or
flag cells as potential mines. The game is won when all non-mine cells are opened and lost when a mine
is opened.
}
\examples{
# To run the Minesweeper game in a local R session:
library(shiny)
sweeper()
}
